fix(migrations): restore CreateBrightSpaceGradeClears registration + code-shape splits
The second-consolidation edit deleted the two folded registrations that
sandwiched CreateBrightSpaceGradeClears and took its (non-folded)
registration with them — every grade-override POST then 500'd with
"no such table: brightspace_grade_clears" (GradeOverridesTests caught it
locally and both CI api-tests lanes confirmed). Registration restored;
a sweep now verifies every file in Migrations/ is registered.

Also lands the deferred code-shape splits, pure moves only:
- PatternFamilyApplication.swift (1,001 lines) -> + PatternFamilyAuthoredGraph.swift
  (cycle detection, authored-ordering reconstruction, section contiguity,
  family-ref dependency validation) + PatternFamilyRendering.swift (the
  deterministic render phase). The applyPatternFamilies function keeps its
  scoped lint disable — decomposing the atomic save path itself is not a
  pure move and stays future work.
- BrightSpaceAPIClient.swift (800 lines) -> BrightSpaceSyncTypes.swift
  (config, error vocabulary, Valence wire DTOs, paged envelope).
- WorkerJobRoutes.swift -> WorkerClaimQueue.swift (the application-level
  claim serializer actor + SQLite-busy retry helpers).

271 tests across the touched suites green; SwiftLint --strict 0
violations across 882 files. The remaining shape items (the
draft-assignment parameter-struct refactor and the assignments.leaf
partial) are signature/template changes, not moves, and are deferred.
