+++ Sources/APIServer/Utilities/PatternFamilyApplication.swift
@@ -685,317 +685,3 @@ func applyPatternFamilies(  // swiftlint:disable:this function_body_length cyclo
         manifestAfter: newManifest
     )
 }
-
-// MARK: - Authored-graph cycle detection
-
-/// Detects dependency cycles on the authored graph where raw scripts are
-/// identified by filename and families are identified by `family:<id>`.
-/// `family:<id>` edges point to the family node (not to its generated
-/// scripts) so the graph stays small.  Uses Kahn's algorithm.
-private func detectAuthoredCycles(
-    authoredRaw: [AuthoredRawScript],
-    families: [PatternFamily]
-) throws {
-    var prereqs: [String: [String]] = [:]  // node → prerequisites of that node
-
-    for r in authoredRaw {
-        let node = r.script
-        prereqs[node, default: []].append(contentsOf: r.dependsOn.map(normaliseNode))
-    }
-    for f in families {
-        let node = familyDepToken(f.id)
-        prereqs[node, default: []].append(contentsOf: f.dependsOn.map(normaliseNode))
-    }
-
-    // Include every referenced prerequisite as a node so Kahn's terminates.
-    for (_, deps) in prereqs {
-        for d in deps where prereqs[d] == nil {
-            prereqs[d] = []
-        }
-    }
-
-    var inDegree: [String: Int] = prereqs.mapValues { $0.count }
-    var dependents: [String: [String]] = [:]
-    for (node, deps) in prereqs {
-        for d in deps {
-            dependents[d, default: []].append(node)
-        }
-    }
-
-    var queue = inDegree.filter { $0.value == 0 }.map(\.key)
-    var processed = 0
-    while !queue.isEmpty {
-        let node = queue.removeLast()
-        processed += 1
-        for dependent in dependents[node, default: []] {
-            inDegree[dependent, default: 0] -= 1
-            if inDegree[dependent] == 0 {
-                queue.append(dependent)
-            }
-        }
-    }
-
-    guard processed == inDegree.count else {
-        throw Abort(
-            .unprocessableEntity,
-            reason: "Dependency graph contains a cycle among scripts and/or pattern families."
-        )
-    }
-}
-
-/// Normalises a dependency token to its canonical node form.  Raw
-/// filenames stay as-is; `family:<id>` tokens keep the prefix so they
-/// don't collide with a real file name like "family" (filename has no
-/// colon; a clash is impossible in practice).
-private func normaliseNode(_ dep: String) -> String { dep }
-
-// MARK: - Extracted phases (#1123)
-
-/// Reconstructs the authored ordering from an existing manifest, for the
-/// legacy (`authoredItems == nil`) path: walk `testSuites` in order, emit a
-/// script item for each raw entry, one family item at the position of each
-/// family's first generated entry, and one check item at the position of
-/// each check's (single) generated entry.  Families/checks present in
-/// `nextFamilies` / `resolvedChecks` but absent from the old manifest
-/// (i.e. newly added) are appended at the end.  This preserves the
-/// instructor's hand-placed position across a family- or check-modal save.
-private func reconstructAuthoredOrdering(
-    props: TestProperties,
-    nextFamilies: [PatternFamily],
-    resolvedChecks: [NotebookCheck],
-    normaliseSectionID: (String?) -> String?
-) -> [AuthoredSuiteItem] {
-    let nextFamilyIDs = Set(nextFamilies.map(\.id))
-    let nextCheckIDs = Set(resolvedChecks.map(\.id))
-    var rebuilt: [AuthoredSuiteItem] = []
-    var seenFamilyIDs: Set<String> = []
-    var seenCheckIDs: Set<String> = []
-    for entry in props.testSuites {
-        if let fid = entry.generatedBy {
-            guard !seenFamilyIDs.contains(fid) else { continue }
-            seenFamilyIDs.insert(fid)
-            if nextFamilyIDs.contains(fid) {
-                rebuilt.append(
-                    .family(
-                        id: fid,
-                        sectionID: normaliseSectionID(entry.sectionID)
-                    ))
-            }
-        } else if let cid = entry.generatedByCheck {
-            guard !seenCheckIDs.contains(cid) else { continue }
-            seenCheckIDs.insert(cid)
-            if nextCheckIDs.contains(cid) {
-                rebuilt.append(
-                    .check(
-                        id: cid,
-                        sectionID: normaliseSectionID(entry.sectionID)
-                    ))
-            }
-        } else {
-            rebuilt.append(
-                .script(
-                    AuthoredRawScript(
-                        script: entry.script,
-                        tier: entry.tier,
-                        points: entry.points,
-                        displayName: entry.name,
-                        dependsOn: entry.dependsOn,
-                        sectionID: normaliseSectionID(entry.sectionID),
-                        hint: entry.hint,
-                        timeLimitSeconds: entry.timeLimitSeconds
-                    )))
-        }
-    }
-    for f in nextFamilies where !seenFamilyIDs.contains(f.id) {
-        rebuilt.append(.family(id: f.id, sectionID: nil))
-    }
-    for c in resolvedChecks where !seenCheckIDs.contains(c.id) {
-        rebuilt.append(.check(id: c.id, sectionID: nil))
-    }
-    return rebuilt
-}
-
-/// Enforces that items with the same `sectionID` form a contiguous block
-/// (nil / ungrouped counts too).  Clients are expected to group `items[]`
-/// before sending; enforcing it server-side catches UI bugs early instead of
-/// producing confusing manifests where the same section straddles another
-/// section.  Internal (not private) so it has direct unit tests.
-func validateAuthoredSectionContiguity(_ items: [AuthoredSuiteItem]) throws {
-    var seenCompleted: Set<String?> = []
-    var current: String?
-    var haveStarted = false
-    for item in items {
-        let sid: String? = {
-            switch item {
-            case .script(let s): return s.sectionID
-            case .family(_, let sid): return sid
-            case .check(_, let sid): return sid
-            }
-        }()
-        if !haveStarted {
-            current = sid
-            haveStarted = true
-            continue
-        }
-        if sid != current {
-            seenCompleted.insert(current)
-            if seenCompleted.contains(sid) {
-                let label = sid ?? "<ungrouped>"
-                throw Abort(
-                    .unprocessableEntity,
-                    reason: "Items with sectionID '\(label)' are not contiguous; "
-                        + "group all items of a section together before saving.")
-            }
-            current = sid
-        }
-    }
-}
-
-/// Validates every `family:<id>` dependency token: raw scripts and notebook
-/// checks may only reference families in `families`; a family may reference
-/// other families but never itself.  (Plain script-filename deps are
-/// validated by `validateManifestDependencies` after expansion, so they
-/// reference an existing entry in the post-expansion `testSuites`.)
-private func validateFamilyRefDependencies(
-    authoredRawEntries: [AuthoredRawScript],
-    families: [PatternFamily],
-    checks: [NotebookCheck]
-) throws {
-    let knownFamilyIDs = Set(families.map(\.id))
-    for r in authoredRawEntries {
-        for dep in r.dependsOn {
-            if let fid = parseFamilyDepToken(dep), !knownFamilyIDs.contains(fid) {
-                throw Abort(
-                    .unprocessableEntity,
-                    reason: "Script '\(r.script)' depends on unknown pattern family '\(fid)'.")
-            }
-        }
-    }
-    for f in families {
-        for dep in f.dependsOn {
-            if let fid = parseFamilyDepToken(dep) {
-                if fid == f.id {
-                    throw Abort(
-                        .unprocessableEntity,
-                        reason: "Pattern family '\(f.id)' cannot depend on itself.")
-                }
-                guard knownFamilyIDs.contains(fid) else {
-                    throw Abort(
-                        .unprocessableEntity,
-                        reason: "Pattern family '\(f.id)' depends on unknown family '\(fid)'.")
-                }
-            }
-        }
-    }
-    for c in checks {
-        for dep in c.dependsOn {
-            if let fid = parseFamilyDepToken(dep), !knownFamilyIDs.contains(fid) {
-                throw Abort(
-                    .unprocessableEntity,
-                    reason: "Notebook check '\(c.id)' depends on unknown pattern family '\(fid)'.")
-            }
-        }
-    }
-}
-
-/// The rendered outputs of every pattern family, keyed by family id —
-/// produced exactly once per apply so the zip write and the manifest rebuild
-/// consume identical artifacts.
-private struct RenderedFamilyArtifacts {
-    /// One rendered script per enabled case, in render order.
-    let caseScripts: [String: [GeneratedScript]]
-    /// The auto existence guard (function-calling kinds only; absent for
-    /// `.variableEquality` / no-enabled-cases families).
-    let guardScripts: [String: GeneratedScript]
-}
-
-/// Renders every family's case scripts and existence guard once.
-private func renderFamilyArtifacts(
-    families: [PatternFamily],
-    familySectionID: [String: String],
-    sectionVarsByID: [String: [FamilyVariable]],
-    globalVariables: [FamilyVariable],
-    perStudentNames: Set<String>,
-    language: AssignmentLanguage = .python
-) -> RenderedFamilyArtifacts {
-    var caseScripts: [String: [GeneratedScript]] = [:]
-    var guardScripts: [String: GeneratedScript] = [:]
-    for family in families {
-        let sectionVariables = familySectionID[family.id].flatMap { sectionVarsByID[$0] } ?? []
-        caseScripts[family.id] = renderPatternFamily(
-            family, sectionVariables: sectionVariables, globalVariables: globalVariables,
-            perStudentNames: perStudentNames, language: language)
-        if let guardScript = existenceGuard(
-            for: family, sectionVariables: sectionVariables, globalVariables: globalVariables,
-            language: language)
-        {
-            guardScripts[family.id] = guardScript
-        }
-    }
-    return RenderedFamilyArtifacts(caseScripts: caseScripts, guardScripts: guardScripts)
-}
-
-/// Slice 1: the re-inline of global + section variables into every raw
-/// (non-generated) Python test script, as a map of zip writes.  Idempotent —
-/// the prepender strips any existing Chickadee inputs block before adding
-/// the new one, so unchanged scripts stay byte-identical and skip the zip
-/// rewrite.  Raw scripts are identified from the authored items (`.script`
-/// case) so the new authored sectionID applies, not whatever the old
-/// manifest had.
-private func rawScriptOverlayWrites(
-    items: [AuthoredSuiteItem],
-    generatedFilenames: Set<String>,
-    zipPath: String,
-    globalVariables: [FamilyVariable],
-    sectionVarsByID: [String: [FamilyVariable]]
-) -> [String: String] {
-    var writes: [String: String] = [:]
-    for item in items {
-        guard case .script(let s) = item else { continue }
-        let filename = s.script
-        // If this filename was just generated (e.g. instructor renamed a
-        // raw script to clash with a family-generated name), the family
-        // version wins — skip the raw-script overlay.
-        guard !generatedFilenames.contains(filename) else { continue }
-        // A raw script's extension names its language: `.py` gets Python
-        // literals, `.R`/`.r` gets R ones. Anything else (a shell script, a
-        // data file) has no literal syntax to inline into and is left alone.
-        let scriptLanguage: AssignmentLanguage?
-        switch (filename as NSString).pathExtension.lowercased() {
-        case "py": scriptLanguage = .python
-        case "r": scriptLanguage = .r
-        default: scriptLanguage = nil
-        }
-        let sectionVars = s.sectionID.flatMap { sectionVarsByID[$0] } ?? []
-        if let provided = s.content {
-            // Declarative content from the payload (the PUT /suite channel
-            // for creating/updating a hand-written script). Write it
-            // verbatim, re-inlining global + section variables for scripts
-            // that have a literal syntax. Idempotent at the zip layer —
-            // identical bytes are a no-op there.
-            writes[filename] =
-                scriptLanguage.map {
-                    TestScriptVariablePrepender.prependToRawScript(
-                        provided, variables: globalVariables + sectionVars, language: $0)
-                } ?? provided
-        } else if let scriptLanguage {
-            // No content provided — preserve the existing file, re-inlining
-            // the current global + section variables (idempotent prepend).
-            guard
-                let existing = readScriptFromZip(
-                    zipPath: zipPath,
-                    filename: filename)
-            else { continue }
-            let updated = TestScriptVariablePrepender.prependToRawScript(
-                existing,
-                variables: globalVariables + sectionVars,
-                language: scriptLanguage
-            )
-            if updated != existing {
-                writes[filename] = updated
-            }
-        }
-        // (.sh and other extensions with no provided content: left untouched.)
-    }
-    return writes
-}
